Collision Pair Namespace Import Reinstate
Bringing this back, as it seems like a lot of the modules appear to have similarly-named items. Tree shaking and or explicit imports will have to come as a result of renaming the imports, or making nicer names for across the entire project. Doing it module by module might not be as maintainable though, as I might accidentally collide names with something else that I will have to use later on. So it will have to happen later if it does get done. Namespaces are okay for the time being though.