Core Engine ESM
I am not always using `const` for seemingly static constant values, as it seems like they sometimes are not actually readonly, so I don't want to break the functionality until it is for certain that it is an intended constant.

Assignment to import object fields will also break when moving to ESM namespaces, as you can't assign to values outside of a module, so that will also have to be accounted for in the event that the library does rely on that. Helper functions are a nice way to go, where the function is the way you pass the value to the module scope. Kind of like a getter/setter in SolidJS.