Default Export Removal
The only one this can't quite be flattened nicely for is the Core Matter module, as that references itself as an object in certain function references. Not totally sure why it is doing that quite yet though, maybe it could be reimplemented without needing that? Will hopefully find more insight about that.